LayoutTests/inspector-protocol/model/content-flow-content-removal.html

<!doctype html>
<html>
<head>
<style>
.content
{
    -webkit-flow-into: flow;
}
</style>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/protocol-test.js"></script>
<script>
function changeFlowContent()
{
    document.getElementById("contentRemoved").remove();
}

function test()
{
    InspectorTest.initializeInspectorModels();

    var contentFlow;

    WebInspector.frameResourceManager.addEventListener(WebInspector.FrameResourceManager.Event.MainFrameDidChange, function(event) {
        var domTree = WebInspector.frameResourceManager.mainFrame.domTree;
        domTree.addEventListener(WebInspector.DOMTree.Event.RootDOMNodeInvalidated, onRootDOMNodeInvalidated, null);
        domTree.addEventListener(WebInspector.DOMTree.Event.ContentFlowWasAdded, onContentFlowWasAdded, null);
        domTree.requestContentFlowList();
    });

    function onRootDOMNodeInvalidated()
    {
        WebInspector.frameResourceManager.mainFrame.domTree.requestContentFlowList();
    }

    function onContentFlowWasAdded(event)
    {
        contentFlow = event.data.flow;
        InspectorTest.assert(contentFlow.name === "flow", "ContentFlow was added.");
        InspectorTest.assert(contentFlow.contentNodes.length === 2, "ContentFlow.contentNodes has a length of 2.");
        InspectorTest.assert(contentFlow.contentNodes[0].getAttribute("id") === "contentStatic", "ContentFlow.contentNodes[0].id is \"#contentStatic\".");
        InspectorTest.assert(contentFlow.contentNodes[1].getAttribute("id") === "contentRemoved", "ContentFlow.contentNodes[1].id is \"#contentRemoved\".");

        contentFlow.addEventListener(WebInspector.ContentFlow.Event.ContentNodeWasRemoved, onContentNodeWasRemoved, null);

        InspectorTest.sendCommand("Runtime.evaluate", {expression: "changeFlowContent()"});
    }

    function onContentNodeWasRemoved(event)
    {
        InspectorTest.assert(event.data.node.getAttribute("id") === "contentRemoved", "\"#contentRemoved\" was removed.");
        InspectorTest.assert(contentFlow.contentNodes.indexOf(event.data.node) === -1, "\"#contentRemoved\" cannot be found in the contentNodes list.");
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
    <p>Testing that the ContentFlows events are correctly dispatched when content nodes are detached from the DOM.</p>

    <div id="contentStatic" class="content"></div>
    <div id="contentRemoved" class="content"></div>

</body>
</html>
